Add files via upload
The SuppressionBot uses datafeed to listen for requests to suppress messages in a pod.  Once the bot user has been added to a private unsearchable administrative room, it can be accessed from that room, or any room where that bot user is added, with this syntax:

   suppress message_id

where message_id is a URL unsafe string obtained by clicking on the timestamp of the message you would like to suppress.  It will be of the format fVyEcE0u3/rDw3AIlhkx6X///qD9oB2gdA==.  Just enter this, and the bot will convert it to a URL safe string, and perform the suppression.

Please note that the Symphony REST API does not offer any mechanism to un-suppress any message.  This action is final and irreversible; additionally it will be captured in the Content Export.

You may also request help from the Suppression Bot like this:

    suppress help