Add files via upload
GreetingBot uses datafeed in rooms where the bot is added as a user to listen to UserAddedToRoom events.  When an event is detected, it will post a message to that room with a proper messageML mention of the new user, so that they know they have been added to the room.  A default message is displayed, or a custom message can be prepared for each room.  The Cmdlets are required for this GreetingBot.